+++ plone/app/discussion/browser/javascripts/controlpanel.js
@@ -100,7 +100,7 @@
         $.updateSettings();
 
         // Set #content class and update settings afterwards
-        $("input,select").live("change", function (e) {
+        $("input,select").on("change", function (e) {
             var id = $(this).attr("id");
             if (id === "form-widgets-globally_enabled-0") {
                 if ($(this).attr("checked")) {