+++ CHANGES.rst
@@ -100,6 +100,8 @@ Changelog
   [toutpt]
 
 
+=======
+>>>>>>> rebased branch from master
 2.2.10 (2013-09-24)
 -------------------